* xdr_reference.c (xdr_reference): Remove cast in argument to memset
	* svc_auth_gssapi.c: In call to gss_accept_sec_context, pass
	OM_uint32 * instead of int *;

	* pmap_clnt.h, pmap_getport.c (pmap_getport): Prototype changed
 	for protocol from unsigned int to rpc_u_int32 to match struct
 	pmap.

	* getrpcport.c (gssrpc_getrpcport): Update to reflect proper
	calling conventions.

	* rpc.h (callrpc): Change prototype from expecting ints for
	prognum, versnum and procnum to rpc_u_int32 for consistancy.
	(getrpcport): Update arguments in similar fashion.

	* clnt_simple.c (gssrpc_callrpc): Update to reflect prototype change.

	* auth_gssapi_misc.c (auth_gssapi_display_status_1): Local
	variable msg_ctx should be OM_uint32 instead of int as argument to
	gss_display_status.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13645 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/rpc/xdr_reference.c
@@ -78,7 +78,7 @@ xdr_reference(xdrs, pp, size, proc)
 				    "xdr_reference: out of memory\n");
 				return (FALSE);
 			}
-			memset(loc, 0, (int)size);
+			memset(loc, 0, size);
 			break;
 
 		case XDR_ENCODE:
